+++ CHANGELOG.md
@@ -2,12 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-<a name="0.0.1"></a>
-## 0.0.1 (2016-11-10)
+<a name="0.0.3"></a>
+## [0.0.3](https://github.com/stormpath/stormpath-sdk-angular/compare/v0.0.1...v0.0.3) (2016-12-02)
 
 
 ### Bug Fixes
 
-* Change to 'inline-source-map' to fix npm test ([eaeceac](https://github.com/stormpath/stormpath-sdk-angular/commit/eaeceac))
-* Fix compilation problems ([c9730ba](https://github.com/stormpath/stormpath-sdk-angular/commit/c9730ba))
-* Polishing ([c45788a](https://github.com/stormpath/stormpath-sdk-angular/commit/c45788a))
+* Change user-agent to be stormpath-sdk-angular ([c42ecc5](https://github.com/stormpath/stormpath-sdk-angular/commit/c42ecc5))
+
+
+
+<a name="0.0.3"></a>
+## [0.0.3](https://github.com/stormpath/stormpath-sdk-angular/compare/v0.0.2...v0.0.3) (2016-12-02)
+
+
+### Features
+
+* Add support for overridding templates ([bc4e6b0](https://github.com/stormpath/stormpath-sdk-angular/commit/bc4e6b0))
+
+<a name="0.0.2"></a>
+## [0.0.2](https://github.com/stormpath/stormpath-sdk-angular/compare/v0.0.1...v0.0.2) (2016-11-11)
+
+### Bug Fixes
+
+* Change user-agent to be stormpath-sdk-angular ([021d882](https://github.com/stormpath/stormpath-sdk-angular/commit/021d882))
+
+<a name="0.0.1"></a>
+## 0.0.1 (2016-11-10)
+
+* Initial Release